feat: draw stats

+++ include/drawing/drawing.h
@@ -5,7 +5,7 @@
 #include <TFT_eSPI.h>
 
 void fillArc(TFT_eSPI &tft, int x, int y, int start_angle, int seg_count, int rx, int ry, int w, unsigned int colour);
-unsigned int brightness(unsigned int colour, int brightness);
-unsigned int rainbow(byte value);
+uint16_t brightness(unsigned int colour, int brightness);
+uint16_t rainbow(byte value);
 
 #endif
